risk-core/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
python-dateutil>=2.8.2
